Improve shell startup history and platform behavior
Changes:

1. Reduce parser and startup storage and preserve fast noninteractive paths.
2. Complete exit diagnostics, AST output, completion, hover, and path rehash behavior.
3. Synchronize custom history through KOSH_HISTORY_FILE and KOSH_HISTORY_SIZE.
4. Harden Windows environment handling and restricted script evaluation.
5. Keep NO_TOILETLINE as an environment-selected build variant.

The shell now uses less compatibility machinery while preserving tested behavior across build variants.

+++ src/Parser.hpp
@@ -84,7 +84,7 @@ private:
 
   Lexer m_lexer;
 
-  usize m_command_depth{0};
+  u16 m_command_depth{0};
   bool m_should_stop_after_top_level_unit{false};
   bool m_has_parsed_source_command{false};
   bool m_should_collect_analysis_metadata{false};
